This is the source code of AdvConstructionSet.
This is my Blitz Basic 2.1 source code of Adventures Construction Set, one of my tries to reproduced a new and better version of the commercial software for Classic Amiga and other platforms that allows to create graphical bidimensional adventures. My very first try was Adventures Conceiver. It requires no additional command in DefLibs. It is not present on Aminet. :(