This is an AmigaOS script to compile with AMOSPro.
This is one of my batch scripts that allows to compile an AMOSPro source using AMOSPro Compiler. I have created it to create the executabled for Adventures Conceiver, seen in normal ways it had no success. It requires a basic knowledge with the compiler options. I suppose that it should works using previous versions of AMOS sources and compiler. Remember that using the INCLIB option let your executable completely standalone so amos.library is no more needed to it!