This is my source code of BlizzardPicture.
This is a source of mine for Blitz Basic 2.1 with the Blitz Support Suite commands addition. It allows the selection of 2 different Workbenches, in my case the OS 3.9 one and the Directory Opus Magellan II other, clicking one of the 2 mouse buttons. Is is devoted to my Amiga 1200 Blizzard 1230 IV accellerator card. You can change the required IFF image and the looped sound by your choiche. Remember that if the program fails no Workbench is loaded!