This is my source code of Frontiera.
I have the classic Amiga game Frontier from Konami. This is my try to obtain a similar but different game lile the commercial one. It has been made by Blitz Basic 2.1 and it requires no additional command but only three sounds. It is present on Aminet without this source code.