This is my source code of Frontiera.
This source code is my poor try to make a game like the Amiga commercial game Frontier that I have purchased so long time ago. It has been made by Blitz Basic 2.1 for the classic Amiga and it requires no additional command in DefLibs. It requires three sounds to run correctly. These are on Aminet with the executable and this source into the Frontiera archive. Enjoy playing if you could! :-)