This is my source code of Frontiera.
This is my poor try to make a game like the commercial Frontier for Classic Amigas using Blitz Basic 2.1. It requires no additional command in DefLibs. This is the version 0.12 of the game and on Aminet you can find the game and its source code at version 0.13 obviously always made by me.