This is my source code of Blizzard Picture.
This is my source code of Blizzard Picture, my first and unique try to launch at boot either the Classic Workbench or the Directory Opus Magellan II Workbench improvement. It is one of my lost source code because when it falis there is no way to launch the Workbench. It requires no additional command in DefLibs. It is not present on Aminet.