This is my source code of AlbaEng.
This is my source code of AlbaEng, the english version of Alba, a program in italian to lock unauthorized access to your Amiga with a keyword. It uses rainbow colours fading backdrop effect that are dfferent on OCS or AGA Amigas. It requires no additional command in DefLibs. It is present on Aminet without this source code.